与netty服务对接

+++ base/base-chat/src/main/java/qsos/base/chat/data/ApiChatMessage.kt
@@ -20,13 +20,13 @@ interface ApiChatMessage {
 
     @POST(value = "$GROUP/sendMessage")
     fun sendMessage(
-            @Header(value = "userId") meId: Int = BaseConfig.userId,
+            @Header(value = "userId") meId: Long = BaseConfig.userId,
             @Body message: ChatMessage
     ): Call<BaseResponse<ChatMessage>>
 
     @GET(value = "$GROUP/getMessageListBySessionIdAndTimeline")
     fun getMessageListBySessionIdAndTimeline(
-            @Header(value = "userId") meId: Int = BaseConfig.userId,
+            @Header(value = "userId") meId: Long = BaseConfig.userId,
             @Query(value = "sessionId") sessionId: Int,
             @Query(value = "timeline") timeline: Int = -1,
             @Query(value = "next") next: Boolean = true,
@@ -35,31 +35,31 @@ interface ApiChatMessage {
 
     @GET(value = "$GROUP/getMessageListBySessionId")
     fun getMessageListBySessionId(
-            @Header(value = "userId") meId: Int = BaseConfig.userId,
+            @Header(value = "userId") meId: Long = BaseConfig.userId,
             @Query(value = "sessionId") sessionId: Int
     ): Call<BaseResponse<List<ChatMessageBo>>>
 
     @POST(value = "$GROUP/readMessage")
     fun readMessage(
-            @Header(value = "userId") meId: Int = BaseConfig.userId,
+            @Header(value = "userId") meId: Long = BaseConfig.userId,
             @Query(value = "messageId") messageId: Int
     ): Call<BaseResponse<ChatMessageReadStatusBo>>
 
     @DELETE(value = "$GROUP/deleteMessage")
     fun deleteMessage(
-            @Header(value = "userId") meId: Int = BaseConfig.userId,
+            @Header(value = "userId") meId: Long = BaseConfig.userId,
             @Query(value = "messageId") messageId: Int
     ): Call<BaseResponse<Boolean>>
 
     @GET(value = "$GROUP/getMessageListByIds")
     fun getMessageListByIds(
-            @Header(value = "userId") meId: Int = BaseConfig.userId,
+            @Header(value = "userId") meId: Long = BaseConfig.userId,
             @Query(value = "messageIds") messageIds: List<Int>
     ): Call<BaseResponse<List<ChatMessageBo>>>
 
     @GET(value = "$GROUP/getMessageById")
     fun getMessageById(
-            @Header(value = "userId") meId: Int = BaseConfig.userId,
+            @Header(value = "userId") meId: Long = BaseConfig.userId,
             @Query(value = "messageId") messageId: Int
     ): Call<BaseResponse<ChatMessageBo>>
