+++ base/base-chat/src/main/java/qsos/base/chat/data/entity/ChatMessageBo.kt
@@ -48,7 +48,7 @@ data class ChatMessageBo(
             message.timeline = value
         }
 
-    override val sendUserId: Int
+    override val sendUserId: Long
         get() = user.userId
 
     override val sendUserName: String
