+++ base/base-chat/src/main/java/qsos/base/chat/view/activity/ChatUserActivity.kt
@@ -29,7 +29,7 @@ class ChatUserActivity(
 
     @Autowired(name = "/CHAT/USER_ID")
     @JvmField
-    var mUserId: Int? = -1
+    var mUserId: Long? = -1L
 
     private var mChatSessionModel: IChatModel.ISession? = null
     private var mChatUserModel: IChatModel.IUser? = null
