+++ base/base-core/src/main/java/qsos/base/core/config/BaseConfig.kt
@@ -5,5 +5,5 @@ package qsos.base.core.config
  * 配置参数
  */
 object BaseConfig {
-    var userId: Int = -1
+    var userId: Long = -1L
 }
